chore(model-integration): move helpers to new package
