+++ model-integration/src/main/java/ai/vespa/modelintegration/utils/package-info.java
@@ -1,5 +1,5 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
-package ai.vespa.utils;
+package ai.vespa.modelintegration.utils;
 
 import com.yahoo.osgi.annotation.ExportPackage;
