+++ model-integration/src/test/java/ai/vespa/embedding/HuggingFaceEmbedderTest.java
@@ -3,6 +3,7 @@ package ai.vespa.embedding;
 
 
 import ai.vespa.modelintegration.evaluator.OnnxRuntime;
+import ai.vespa.modelintegration.utils.ModelPathHelper;
 import com.yahoo.config.ModelReference;
 import com.yahoo.embedding.huggingface.HuggingFaceEmbedderConfig;
 import com.yahoo.language.process.Embedder;
