+++ model-integration/src/test/java/ai/vespa/modelintegration/utils/ModelPathHelperImplTest.java
@@ -1,6 +1,6 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-package ai.vespa.embedding;
+package ai.vespa.modelintegration.utils;
 
 import ai.vespa.secret.Secret;
 import ai.vespa.secret.Secrets;
