+++ model-integration/src/test/java/ai/vespa/modelintegration/utils/OnnxExternalDataResolverTest.java
@@ -1,10 +1,7 @@
-/*
- * Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
- */
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-package ai.vespa.utils;
+package ai.vespa.modelintegration.utils;
 
-import ai.vespa.embedding.ModelPathHelper;
 import com.yahoo.config.ModelReference;
 import com.yahoo.config.UrlReference;
 import org.junit.jupiter.api.Test;
